+++ src/SMAPI/Framework/SMultiplayer.cs
@@ -196,7 +196,13 @@ namespace StardewModdingAPI.Framework
                         this.Monitor.Log($"Received context for farmhand {message.FarmerID} running {(model != null ? $"SMAPI {model.ApiVersion} with {model.Mods.Length} mods" : "vanilla")}.", LogLevel.Trace);
 
                         // store peer
-                        MultiplayerPeer newPeer = new MultiplayerPeer(message.FarmerID, model, sendMessage, isHost: false);
+                        MultiplayerPeer newPeer = new MultiplayerPeer(
+                            playerID: message.FarmerID,
+                            screenID: this.GetScreenId(message.FarmerID),
+                            model: model,
+                            sendMessage: sendMessage,
+                            isHost: false
+                        );
                         if (this.Peers.ContainsKey(message.FarmerID))
                         {
                             this.Monitor.Log($"Received mod context from farmhand {message.FarmerID}, but the game didn't see them disconnect. This may indicate issues with the network connection.", LogLevel.Info);
@@ -238,7 +244,13 @@ namespace StardewModdingAPI.Framework
                     if (!this.Peers.ContainsKey(message.FarmerID))
                     {
                         this.Monitor.Log($"Received connection for vanilla player {message.FarmerID}.", LogLevel.Trace);
-                        MultiplayerPeer peer = new MultiplayerPeer(message.FarmerID, null, sendMessage, isHost: false);
+                        MultiplayerPeer peer = new MultiplayerPeer(
+                            playerID: message.FarmerID,
+                            screenID: this.GetScreenId(message.FarmerID),
+                            model: null,
+                            sendMessage: sendMessage,
+                            isHost: false
+                        );
                         this.AddPeer(peer, canBeHost: false);
                     }
 
@@ -280,7 +292,13 @@ namespace StardewModdingAPI.Framework
                         this.Monitor.Log($"Received context for {(model?.IsHost == true ? "host" : "farmhand")} {message.FarmerID} running {(model != null ? $"SMAPI {model.ApiVersion} with {model.Mods.Length} mods" : "vanilla")}.", LogLevel.Trace);
 
                         // store peer
-                        MultiplayerPeer peer = new MultiplayerPeer(message.FarmerID, model, sendMessage, isHost: model?.IsHost ?? this.HostPeer == null);
+                        MultiplayerPeer peer = new MultiplayerPeer(
+                            playerID: message.FarmerID,
+                            screenID: this.GetScreenId(message.FarmerID),
+                            model: model,
+                            sendMessage: sendMessage,
+                            isHost: model?.IsHost ?? this.HostPeer == null
+                        );
                         if (peer.IsHost && this.HostPeer != null)
                         {
                             this.Monitor.Log($"Rejected mod context from host player {peer.PlayerID}: already received host data from {(peer.PlayerID == this.HostPeer.PlayerID ? "that player" : $"player {peer.PlayerID}")}.", LogLevel.Error);
@@ -297,7 +315,14 @@ namespace StardewModdingAPI.Framework
                         if (!this.Peers.ContainsKey(message.FarmerID) && this.HostPeer == null)
                         {
                             this.Monitor.Log($"Received connection for vanilla host {message.FarmerID}.", LogLevel.Trace);
-                            this.AddPeer(new MultiplayerPeer(message.FarmerID, null, sendMessage, isHost: true), canBeHost: false);
+                            var peer = new MultiplayerPeer(
+                                playerID: message.FarmerID,
+                                screenID: this.GetScreenId(message.FarmerID),
+                                model: null,
+                                sendMessage: sendMessage,
+                                isHost: true
+                            );
+                            this.AddPeer(peer, canBeHost: false);
                         }
                         resume();
                         break;
@@ -309,7 +334,13 @@ namespace StardewModdingAPI.Framework
                         // store peer
                         if (!this.Peers.TryGetValue(message.FarmerID, out MultiplayerPeer peer))
                         {
-                            peer = new MultiplayerPeer(message.FarmerID, null, sendMessage, isHost: this.HostPeer == null);
+                            peer = new MultiplayerPeer(
+                                playerID: message.FarmerID,
+                                screenID: this.GetScreenId(message.FarmerID),
+                                model: null,
+                                sendMessage: sendMessage,
+                                isHost: this.HostPeer == null
+                            );
                             this.Monitor.Log($"Received connection for vanilla {(peer.IsHost ? "host" : "farmhand")} {message.FarmerID}.", LogLevel.Trace);
                             this.AddPeer(peer, canBeHost: true);
                         }
@@ -503,6 +534,13 @@ namespace StardewModdingAPI.Framework
                     }
                 }
             }
+        }
+
+        /// <summary>Get the screen ID for a given player ID, if the player is local.</summary>
+        /// <param name="playerId">The player ID to check.</param>
+        private int? GetScreenId(long playerId)
+        {
+            return SGameRunner.Instance.GetScreenId(playerId);
         }
 
         /// <summary>Get all connected player IDs, including the current player.</summary>